address codec clang-tidy fix
