+++ core/primitives/address/address.hpp
@@ -5,7 +5,6 @@
 
 #pragma once
 
-#include <spdlog/fmt/fmt.h>
 #include <boost/variant.hpp>
 
 #include "common/blob.hpp"
@@ -107,19 +106,7 @@ namespace fc::primitives::address {
    * @brief Addresses "less than" operator
    */
   bool operator<(const Address &lhs, const Address &rhs);
-
-  std::string encodeToString(const Address &address);
 }  // namespace fc::primitives::address
-
-template <>
-struct fmt::formatter<fc::primitives::address::Address>
-    : formatter<std::string_view> {
-  template <typename C>
-  auto format(const fc::primitives::address::Address &address, C &ctx) {
-    auto str = fc::primitives::address::encodeToString(address);
-    return formatter<std::string_view>::format(str, ctx);
-  }
-};
 
 /**
  * @brief Outcome errors declaration
